requirements.txt
aiofiles==23.2.1
aiohttp==3.8.6
aiosignal==1.3.1
alembic==1.12.1
altair==5.1.2
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
anyio==3.7.1
asteroid-filterbanks==0.4.0
async-timeout==4.0.3
attrs==23.1.0
audioread==3.0.1
beautifulsoup4==4.12.2
blinker==1.7.0
Brotli==1.1.0
cachetools==5.3.2
certifi==2022.12.7
cffi==1.16.0
charset-normalizer==2.1.1
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
colorlog==6.7.0
contourpy==1.2.0
cycler==0.12.1
decorator==4.4.2
distro==1.8.0
docopt==0.6.2
einops==0.7.0
entrypoints==0.4
Faker==20.0.0
favicon==0.7.0
filelock==3.9.0
flatbuffers==23.5.26
fonttools==4.44.0
frozenlist==1.4.0
fsspec==2023.10.0
gitdb==4.0.11
GitPython==3.1.40
GPUtil==1.4.0
greenlet==3.0.1
h11==0.14.0
htbuilder==0.6.2
httpcore==1.0.2
httpx==0.25.1
huggingface-hub==0.19.0
humanfriendly==10.0
HyperPyYAML==1.2.2
idna==3.4
imageio==2.32.0
imageio-ffmpeg==0.4.9
importlib-metadata==6.8.0
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
julius==0.2.7
kiwisolver==1.4.5
langcodes==3.3.0
language-data==1.1
lazy_loader==0.3
librosa==0.10.1
lightning==2.1.1
lightning-utilities==0.9.0
llvmlite==0.41.1
lxml==4.9.3
Mako==1.3.0
marisa-trie==0.7.8
Markdown==3.5.1
markdown-it-py==3.0.0
markdownlit==0.0.7
MarkupSafe==2.1.3
matplotlib==3.8.1
mdurl==0.1.2
more-itertools==10.1.0
moviepy==1.0.3
mpmath==1.3.0
msgpack==1.0.7
multidict==6.0.4
mutagen==1.47.0
networkx==3.0
numba==0.58.1
numpy==1.26.2
omegaconf==2.3.0
onnxruntime==1.16.2
onnxruntime-gpu==1.16.2
openai==1.2.3
openai-whisper==20231106
optuna==3.4.0
packaging==23.2
pandas==2.1.3
Pillow==9.3.0
platformdirs==4.0.0
pooch==1.8.0
primePy==1.3
proglog==0.1.10
protobuf==4.25.0
pyannote.audio==3.0.1
pyannote.core==5.0.0
pyannote.database==5.0.1
pyannote.metrics==3.2.1
pyannote.pipeline==3.0.1
pyarrow==14.0.1
pycparser==2.21
pycryptodomex==3.19.0
pydantic==2.4.2
pydantic_core==2.10.1
pydeck==0.8.1b0
pydub==0.25.1
Pygments==2.16.1
pymdown-extensions==10.4
pyparsing==3.1.1
pyperclip==1.8.2
pyreadline3==3.4.1
python-dateutil==2.8.2
pytorch-lightning==2.1.1
pytorch-metric-learning==2.3.0
pytube==15.0.0
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.30.2
regex==2023.10.3
requests==2.28.1
rich==13.6.0
rpds-py==0.12.0
ruamel.yaml==0.18.5
ruamel.yaml.clib==0.2.8
scikit-learn==1.3.2
scipy==1.11.3
semver==3.0.2
sentencepiece==0.1.99
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
sortedcontainers==2.4.0
soundfile==0.12.1
soupsieve==2.5
soxr==0.3.7
speechbrain==0.5.15
SQLAlchemy==2.0.23
st-annotated-text==4.0.1
st-pages==0.4.5
streamlit==1.28.1
streamlit-antd-components==0.2.3
streamlit-camera-input-live==0.2.0
streamlit-card==0.0.61
streamlit-embedcode==0.1.2
streamlit-extras==0.3.5
streamlit-faker==0.0.3
streamlit-image-coordinates==0.1.6
streamlit-keyup==0.2.0
streamlit-player==0.1.5
streamlit-toggle-switch==1.0.2
streamlit-vertical-slider==1.0.2
sympy==1.12
tabulate==0.9.0
tenacity==8.2.3
tensorboardX==2.6.2.2
threadpoolctl==3.2.0
tiktoken==0.5.1
toml==0.10.2
toolz==0.12.0
torch==2.0.0+cu118
torch-audiomentations==0.11.0
torch-pitch-shift==1.2.4
torchaudio==2.0.1+cu118
torchmetrics==1.2.0
torchvision==0.15.1+cu118
tornado==6.3.3
tqdm==4.66.1
typer==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.2
urllib3==1.26.13
validators==0.22.0
watchdog==3.0.0
websockets==12.0
yarl==1.9.2
yt-dlp==2023.10.13
zipp==3.17.0
